+++ tests_compilations/calculus_test.py
@@ -20,6 +20,7 @@
 # Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 import subprocess
+from pathlib import Path
 
 from mathmakerlib.calculus import Table
 from .compilation_manager import CompilationManager
@@ -36,4 +37,28 @@ def test1():
         # sys.stderr.write(f'cmd={cmd}\n')
         ret_code = subprocess.run(cmd, shell=True, executable='/bin/bash')\
             .returncode
+        assert ret_code == 0
+
+
+def test_PythagoreanEquation_compilations():
+    DATA_PATH = Path(__file__).parent / 'data/pythagorean_equations'
+    TEST_ABC_1 = (DATA_PATH / 'ABC_1.tex').read_text()
+    TEST_ABC_2 = (DATA_PATH / 'ABC_2.tex').read_text()
+    TEST_GIH = (DATA_PATH / 'GIH.tex').read_text()
+    TEST_GMW = (DATA_PATH / 'GMW.tex').read_text()
+    TEST_ZIP = (DATA_PATH / 'ZIP.tex').read_text()
+    TEST_ZIP1 = (DATA_PATH / 'ZIP1.tex').read_text()
+    TEST_ZIP2 = (DATA_PATH / 'ZIP2.tex').read_text()
+    TEST_ZIP3 = (DATA_PATH / 'ZIP3.tex').read_text()
+    TEST_SVK = (DATA_PATH / 'SVK.tex').read_text()
+    TEST_SVK1 = (DATA_PATH / 'SVK1.tex').read_text()
+    TEST_SVK2 = (DATA_PATH / 'SVK2.tex').read_text()
+    TEST_SVK3 = (DATA_PATH / 'SVK3.tex').read_text()
+    content = '\n'.join([TEST_ABC_1, TEST_ABC_2, TEST_GIH, TEST_GMW, TEST_ZIP,
+                        TEST_ZIP1, TEST_ZIP2, TEST_ZIP3, TEST_SVK, TEST_SVK1,
+                        TEST_SVK2, TEST_SVK3])
+    with CompilationManager('test_PythagoreanEquation_compilations',
+                            'article.tex', content) as cmd:
+        ret_code = subprocess.run(cmd, shell=True, executable='/bin/bash',
+                                  capture_output=True).returncode
         assert ret_code == 0